[IMP] Moved is_user_authenticated into namespace blueprints
